Add Entry domain model and refactor waiter logic
Introduced the Entry class to the domain models and included it in the project file. Removed Mozo_Fecha from the Waiter class and commented out related logic for cleaning inactive waiters in AuthHelper and MasterPage. Also removed the reset of active waiters on application start in Global.asax. Minor variable name correction in Login.aspx.cs.